Hint at `tool.uv.environments` on resolution error
Users are not (yet) properly familiar with the concept of universal resolution and its implication that we need to resolve for all possible platforms and Python versions. Some projects only target a specific platform or Python version and experience resolution errors due to failures for other platforms. Indicated by the number of questions we get about it, `too.uv.environments` for restricting environments is not well discoverable.

We add a special hint when resolution failed on a fork disjoint with the current environment, hinting the user to constraint `requires-python` and `tool.uv.environments` respectively.

The hint has false positives for cases where the resolution failed on a different platform, but equally fails on the current platform, but the other fork was earlier. Given that conflicts can be based on `requires-python`, afaik be can't parse whether the current platform would also be affected from the derivation tree.

Two cases not covered by this are build errors as well as install error that need `tool.uv.required-environments`.